Use `pushadd_to_gateway` instead of `push_to_gateway`
Only the last job gets recorded right now, and I believe that is because
we are calling push_to_gateway multiple times. This is just a quick fix.

The real fix would be to add metrics to the registry and not push to the
pushgateway until the pytest session finished hook triggers (not sure of
the exact name)